DecksController Authrization

+++ PowerCards/Controllers/DecksController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http.Features;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using PowerCards.DAL;
@@ -15,11 +16,13 @@ namespace PowerCards.Controllers
     {
         private readonly IDeckRepository _deckRepository;
         private readonly ILogger<DecksController> _logger;
+        private readonly UserManager<User> _userManager;
 
-        public DecksController(IDeckRepository deckRepository, ILogger<DecksController> logger)
+        public DecksController(IDeckRepository deckRepository, ILogger<DecksController> logger,  UserManager<User> userManager)
         {
             _deckRepository = deckRepository;
             _logger = logger;
+            _userManager = userManager;
         }
 
         public async Task<IActionResult> Index()
@@ -88,7 +91,14 @@ namespace PowerCards.Controllers
                 _logger.LogError("[DeckController] Deck not found when updating/editing in the DeckID {DeckID:0000", id);
                 return BadRequest("Deck not found for the DeckID");
             }
+            var currUsername = User.Identity.Name;
+            if (deck.UserName != currUsername)
+            {
+                return Unauthorized("You are not the user");
+            }   
             return View(deck);
+            
+           
 
         }
 
@@ -96,14 +106,22 @@ namespace PowerCards.Controllers
         [Authorize]
         public async Task<IActionResult> Edit(Deck deck)
         {
+
             if (ModelState.IsValid)
             {
                 bool success = await _deckRepository.Edit(deck);
                 if (success)
                     return RedirectToAction(nameof(Index));
-            }
+            } 
             _logger.LogWarning("[DecksController] Failed to edit deck with DeckID: {DeckID}", deck.DeckID);
+            var currUsername = User.Identity.Name;
+            if (deck.UserName != currUsername)
+            {
+                return Unauthorized("You are not the user");
+            }
             return View(deck);
+        
+
         }
 
         [HttpGet]
@@ -116,6 +134,11 @@ namespace PowerCards.Controllers
                 _logger.LogError("[DeckController] Deck not found for the DeckID {DeckID: 0000}", id);
                 return BadRequest("Deck not found for the DeckID");
             }
+            var currUsername = User.Identity.Name;
+            if (deck.UserName != currUsername)
+            {
+                return Unauthorized("You are not the user");
+            }
             return View(deck);
 
         }
@@ -124,12 +147,24 @@ namespace PowerCards.Controllers
         [Authorize]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var DeckDelete = await _deckRepository.GetById(id);
+            if (DeckDelete == null)
+            {
+                _logger.LogError("[DecksController] Deck not found for the DeckID {DeckID: 0000}", id);
+                return BadRequest("Deck not found for the DeckID");
+            }
+            var currUsername = User.Identity.Name;
+            if(DeckDelete.UserName != currUsername)
+            {
+                return Unauthorized("You are not the user");
+            }
             var success = await _deckRepository.Delete(id);
             if (!success)
             {
                 _logger.LogError("[DecksController] Failed to delete deck with DeckID: {DeckID}", id);
                 return NotFound();
             }
+
             return RedirectToAction(nameof(Index));
         }
     }
